requirements.txt
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
openai>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0 